Add files via upload
In multi-path inheritance , if we observe carefully, we find that this type of inheritance is actually the combination of three different types of inheritance- Hierarchical,multiple and Multi_level inheritance.
The derived class inherits the members of the base class(Grandparent) twice, through parent_1(Derived class_1) and parent class_2(Derived class_2). This results in ambiguity because a duplicate set of members is created. This type of Ambiguous situation must be avoided.